# api/openapi/trainings.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Wolf Workouts - Trainings API
  description: API for trainings management
  contact:
    name: David Alecrim
    url: https://github.com/davidalecrim1
servers:
  - url: https://{hostname}/api
    variables:
      hostname:
        default: localhost
security:
  - bearerAuth: []
paths:
  /trainings:
    post:
      operationId: scheduleTraining
      summary: Schedule a new training
      description: Schedules a new training
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleTrainingRequest"
      responses:
        "201":
          description: Training scheduled successfully
        "400":
          description: Invalid request body
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
    get:
      operationId: getTrainings
      summary: Get all trainings
      description: Retrieves all trainings for the authenticated user
      responses:
        "200":
          description: Trainings retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetTrainingsResponse"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseError"
components:
  schemas:
    ScheduleTrainingRequest:
      type: object
      properties:
        training_datetime:
          type: string
          format: date-time
        notes:
          type: string
      required:
        - training_datetime
        - notes
    ResponseError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    GetTrainingsResponse:
      type: array
      items:
        $ref: "#/components/schemas/Training"
    Training:
      type: object
      properties:
        id:
          type: string
        notes:
          type: string
        training_datetime:
          type: string
          format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT