# internal/trainings/handler/generated/openapi_types.gen.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.2.0 DO NOT EDIT.
package generated

import (
	"time"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// GetTrainingsResponse defines model for GetTrainingsResponse.
type GetTrainingsResponse = []Training

// ResponseError defines model for ResponseError.
type ResponseError struct {
	Message string `json:"message"`
}

// ScheduleTrainingRequest defines model for ScheduleTrainingRequest.
type ScheduleTrainingRequest struct {
	Notes            string    `json:"notes"`
	TrainingDatetime time.Time `json:"training_datetime"`
}

// Training defines model for Training.
type Training struct {
	Id               *string    `json:"id,omitempty"`
	Notes            *string    `json:"notes,omitempty"`
	TrainingDatetime *time.Time `json:"training_datetime,omitempty"`
}

// ScheduleTrainingJSONRequestBody defines body for ScheduleTraining for application/json ContentType.
type ScheduleTrainingJSONRequestBody = ScheduleTrainingRequest